Split definitions into reusable module
