+++ Assets/Lua/Doom/things-lines.lua
@@ -1,5 +1,7 @@
 -- feos, 2025
 
+local dsda = require("dsda-data")
+
 -- CONSTANTS
 local NULL_OBJECT      = 0x88888888 -- no object at that index
 local OUT_OF_BOUNDS    = 0xFFFFFFFF -- no such index
@@ -10,11 +12,6 @@ local CHAR_WIDTH       = 10
 local CHAR_HEIGHT      = 16
 local NEGATIVE_MAXIMUM = 1 << 63
 local POSITIVE_MAXIMUM = ~NEGATIVE_MAXIMUM
-local MAX_PLAYERS      = 4
--- sizes in bytes
-local LINE_SIZE   = 256  -- sizeof(line_t) is 232, but we padded it for niceness
-local MOBJ_SIZE   = 512  -- sizeof(mobj_t) is 464, but we padded it for niceness
-local PLAYER_SIZE = 1024 -- sizeof(player_t) is 729, but we padded it for niceness
 -- shortcuts
 local rl   = memory.read_u32_le
 local rw   = memory.read_u16_le
@@ -48,10 +45,10 @@ local LastScreenSize = {
 	h = client.screenheight()
 }
 -- forward declarations
-local PlayerOffsets = {}-- player member offsets in bytes
-local MobjOffsets   = {} -- mobj member offsets in bytes
-local MobjType      = {}
-local SpriteNumber  = {}
+local PlayerOffsets = dsda.player.offsets-- player member offsets in bytes
+local MobjOffsets   = dsda.mobj.offsets-- mobj member offsets in bytes
+local MobjType      = dsda.doom.mobjtype
+local SpriteNumber  = dsda.doom.spritenum
 local Objects       = {}
 
 --gui.defaultPixelFont("fceux")
@@ -133,8 +130,8 @@ local function iterate_players()
 	local total_itemcount   = 0
 	local total_secretcount = 0
 	local stats             = "      HP Armr Kill Item Secr\n"
-	for i = 1, MAX_PLAYERS do
-		local addr = PLAYER_SIZE * (i - 1)
+	for i = 1, dsda.MAX_PLAYERS do
+		local addr = dsda.PLAYER_SIZE * (i - 1)
 		local mobj = rl(addr + PlayerOffsets.mobj, "Players")
 
 		if mobj ~= NULL_OBJECT then
@@ -191,7 +188,7 @@ local function iterate()
 	end
 	
 	for i = 0, 100000 do
-		local addr = i * LINE_SIZE
+		local addr = i * dsda.LINE_SIZE
 		if addr > 0xFFFFFF then break end
 		
 		local id = rl(addr, "Lines") & 0xFFFFFFFF
@@ -215,7 +212,7 @@ end
 
 local function init_objects()
 	for i = 0, 100000 do
-		local addr = i * MOBJ_SIZE
+		local addr = i * dsda.MOBJ_SIZE
 		if addr > 0xFFFFFF then break end
 		
 		local thinker = rl(addr, "Things") & 0xFFFFFFFF -- just to check if mobj is there
@@ -299,591 +296,7 @@ local function make_button(x, y, name, func)
 	text(textX, textY, name, colors[colorIndex] | 0xff000000) -- full alpha
 end
 
-local function struct_layout(struct)
-	struct = struct or {}
-	struct.size = 0
-	struct.offsets = {}
 
-	function struct.add(name, size, alignment)
-		if alignment == true then alignment = size end
-		struct.align(alignment)
-		--print(string.format("%-19s %3X %3X", name, size, struct.size)); emu.yield()
-		struct.offsets[name] = struct.size
-		struct.size = struct.size + size
-		struct.align(alignment) -- add padding to structs
-		return struct
-	end
-	function struct.align(alignment)
-		if alignment and struct.size % alignment > 0 then
-			--print(string.format("%i bytes padding", alignment - (struct.size % alignment)))
-			struct.pad(alignment - (struct.size % alignment))
-		end
-	end
-	function struct.pad(size)
-		struct.size = struct.size + size
-		return struct
-	 end
-	function struct.s8   (name) return struct.add(name, 1, true) end
-	function struct.s16  (name) return struct.add(name, 2, true) end
-	function struct.s32  (name) return struct.add(name, 4, true) end
-	function struct.u8   (name) return struct.add(name, 1, true) end
-	function struct.u32  (name) return struct.add(name, 4, true) end
-	function struct.u64  (name) return struct.add(name, 8, true) end
-	function struct.ptr  (name) return struct.u64(name) end
-	function struct.bool (name) return struct.s32(name) end
-	function struct.array(name, type, count, ...)
-		for i = 1, count do
-			struct[type](name .. i, ...)
-		end
-		return struct
-	end
-
-	return struct
-end
-
---[[
-player_t https://github.com/TASEmulators/dsda-doom/blob/master/prboom2/src/d_player.h
-mobj              8   0
-playerstate       4   8
-cmd               E   C
-viewz             4  1C
-viewheight        4  20
-deltaviewheight   4  24
-bob               4  28
-health            4  2C
-armorpoints1      4  30
-armorpoints2      4  34
-armorpoints3      4  38
-armorpoints4      4  3C
-armortype         4  40
-powers1           4  44
-powers2           4  48
-powers3           4  4C
-powers4           4  50
-powers5           4  54
-powers6           4  58
-powers7           4  5C
-powers8           4  60
-powers9           4  64
-powers10          4  68
-powers11          4  6C
-powers12          4  70
-cards1            4  74
-cards2            4  78
-cards3            4  7C
-cards4            4  80
-cards5            4  84
-cards6            4  88
-cards7            4  8C
-cards8            4  90
-cards9            4  94
-cards10           4  98
-cards11           4  9C
-backpack          4  A0
-frags1            4  A4
-frags2            4  A8
-frags3            4  AC
-frags4            4  B0
-frags5            4  B4
-frags6            4  B8
-frags7            4  BC
-frags8            4  C0
-readyweapon       4  C4
-pendingweapon     4  C8
-weaponowned1      4  CC
-weaponowned2      4  D0
-weaponowned3      4  D4
-weaponowned4      4  D8
-weaponowned5      4  DC
-weaponowned6      4  E0
-weaponowned7      4  E4
-weaponowned8      4  E8
-weaponowned9      4  EC
-ammo1             4  F0
-ammo2             4  F4
-ammo3             4  F8
-ammo4             4  FC
-ammo5             4 100
-ammo6             4 104
-maxammo1          4 108
-maxammo2          4 10C
-maxammo3          4 110
-maxammo4          4 114
-maxammo5          4 118
-maxammo6          4 11C
-attackdown        4 120
-usedown           4 124
-cheats            4 128
-refire            4 12C
-killcount         4 130
-itemcount         4 134
-secretcount       4 138
-damagecount       4 13C
-bonuscount        4 140
-attacker          8 148
-extralight        4 150
-fixedcolormap     4 154
-colormap          4 158
-psprites         30 160
-didsecret         4 190
-momx              4 194
-mony              4 198
-maxkilldiscount   4 19C
-prev_viewz        4 1A0
-prev_viewangle    4 1A4
-prev_viewpitch    4 1A8
-(padding omitted)
-]]
-
-PlayerOffsets = struct_layout()
-	.ptr  ("mobj")
-	.s32  ("playerstate")
-	.add  ("cmd", 14, 2)
-	.s32  ("viewz")
-	.s32  ("viewheight")
-	.s32  ("deltaviewheight")
-	.s32  ("bob")
-	.s32  ("health")
-	.array("armorpoints", "s32", 4)
-	.s32  ("armortype")
-	.array("powers", "s32", 12)
-	.array("cards", "bool", 11)
-	.bool ("backpack")
-	.array("frags", "s32", 8)
-	.s32  ("readyweapon")
-	.s32  ("pendingweapon")
-	.array("weaponowned", "bool", 9)
-	.array("ammo", "s32", 6)
-	.array("maxammo", "s32", 6)
-	.s32  ("attackdown")
-	.s32  ("usedown")
-	.s32  ("cheats")
-	.s32  ("refire")
-	.s32  ("killcount")
-	.s32  ("itemcount")
-	.s32  ("secretcount")
-	.s32  ("damagecount")
-	.s32  ("bonuscount")
-	.ptr  ("attacker")
-	.s32  ("extralight")
-	.s32  ("fixedcolormap")
-	.s32  ("colormap")
-	.add  ("psprites", 24*2, 8)
-	.bool ("didsecret")
-	.s32  ("momx")
-	.s32  ("mony")
-	.s32  ("maxkilldiscount")
-	.s32  ("prev_viewz")
-	.u32  ("prev_viewangle")
-	.u32  ("prev_viewpitch")
-	-- the rest are non-doom
-	.offsets
-
---[[
-mobj_t https://github.com/TASEmulators/dsda-doom/blob/master/prboom2/src/p_mobj.h
-thinker              2C   0
-x                     4  30
-y                     4  34
-z                     4  38
-snext                 8  40
-sprev                 8  48
-angle                 4  50
-sprite                4  54
-frame                 4  58
-bnext                 8  60
-bprev                 8  68
-subsector             8  70
-floorz                4  78
-ceilingz              4  7C
-dropoffz              4  80
-radius                4  84
-height                4  88
-momx                  4  8C
-momy                  4  90
-momz                  4  94
-validcount            4  98
-type                  4  9C
-info                  8  A0
-tics                  4  A8
-state                 8  B0
-flags                 8  B8
-intflags              4  C0
-health                4  C4
-movedir               2  C8
-movecount             2  CA
-strafecount           2  CC
-target                8  D0
-reactiontime          2  D8
-threshold             2  DA
-pursuecount           2  DC
-gear                  2  DE
-player                8  E0
-lastlook              2  E8
-spawnpoint           3A  EC
-tracer                8 128
-lastenemy             8 130
-friction              4 138
-movefactor            4 13C
-touching_sectorlist   8 140
-PrevX                 4 148
-PrevY                 4 14C
-PrevZ                 4 150
-pitch                 4 154
-index                 4 158
-patch_width           2 15C
-iden_nums             4 160
-(padding omitted)
---]]--
-
-MobjOffsets = struct_layout()
-	.add("thinker", 44, 8)
-	.s32("x")
-	.s32("y")
-	.s32("z")
-	.ptr("snext")
-	.ptr("sprev")
-	.u32("angle")
-	.s32("sprite")
-	.s32("frame")
-	.ptr("bnext")
-	.ptr("bprev")
-	.ptr("subsector")
-	.s32("floorz")
-	.s32("ceilingz")
-	.s32("dropoffz")
-	.s32("radius")
-	.s32("height")
-	.s32("momx")
-	.s32("momy")
-	.s32("momz")
-	.s32("validcount")
-	.s32("type")
-	.ptr("info")
-	.s32("tics")
-	.ptr("state")
-	.u64("flags")
-	.s32("intflags")
-	.s32("health")
-	.s16("movedir")
-	.s16("movecount")
-	.s16("strafecount")
-	.ptr("target")
-	.s16("reactiontime")
-	.s16("threshold")
-	.s16("pursuecount")
-	.s16("gear")
-	.ptr("player")
-	.s16("lastlook")
-	.add("spawnpoint", 58, 4)
-	.ptr("tracer")
-	.ptr("lastenemy")
-	.s32("friction")
-	.s32("movefactor")
-	.ptr("touching_sectorlist")
-	.s32("PrevX")
-	.s32("PrevY")
-	.s32("PrevZ")
-	.u32("pitch")
-	.s32("index")
-	.s16("patch_width")
-	.s32("iden_nums")
-	-- the rest are non-doom
-	.offsets
-
-MobjType = {
---	"NULL" = -1,
---	"ZERO",
---	"PLAYER = ZERO",
-	"POSSESSED",
-	"SHOTGUY",
-	"VILE",
-	"FIRE",
-	"UNDEAD",
-	"TRACER",
-	"SMOKE",
-	"FATSO",
-	"FATSHOT",
-	"CHAINGUY",
-	"TROOP",
-	"SERGEANT",
-	"SHADOWS",
-	"HEAD",
-	"BRUISER",
-	"BRUISERSHOT",
-	"KNIGHT",
-	"SKULL",
-	"SPIDER",
-	"BABY",
-	"CYBORG",
-	"PAIN",
-	"WOLFSS",
-	"KEEN",
-	"BOSSBRAIN",
-	"BOSSSPIT",
-	"BOSSTARGET",
-	"SPAWNSHOT",
-	"SPAWNFIRE",
-	"BARREL",
-	"TROOPSHOT",
-	"HEADSHOT",
-	"ROCKET",
-	"PLASMA",
-	"BFG",
-	"ARACHPLAZ",
-	"PUFF",
-	"BLOOD",
-	"TFOG",
-	"IFOG",
-	"TELEPORTMAN",
-	"EXTRABFG",
-	"MISC0",
-	"MISC1",
-	"MISC2",
-	"MISC3",
-	"MISC4",
-	"MISC5",
-	"MISC6",
-	"MISC7",
-	"MISC8",
-	"MISC9",
-	"MISC10",
-	"MISC11",
-	"MISC12",
-	"INV",
-	"MISC13",
-	"INS",
-	"MISC14",
-	"MISC15",
-	"MISC16",
-	"MEGA",
-	"CLIP",
-	"MISC17",
-	"MISC18",
-	"MISC19",
-	"MISC20",
-	"MISC21",
-	"MISC22",
-	"MISC23",
-	"MISC24",
-	"MISC25",
-	"CHAINGUN",
-	"MISC26",
-	"MISC27",
-	"MISC28",
-	"SHOTGUN",
-	"SUPERSHOTGUN",
-	"MISC29",
-	"MISC30",
-	"MISC31",
-	"MISC32",
-	"MISC33",
-	"MISC34",
-	"MISC35",
-	"MISC36",
-	"MISC37",
-	"MISC38",
-	"MISC39",
-	"MISC40",
-	"MISC41",
-	"MISC42",
-	"MISC43",
-	"MISC44",
-	"MISC45",
-	"MISC46",
-	"MISC47",
-	"MISC48",
-	"MISC49",
-	"MISC50",
-	"MISC51",
-	"MISC52",
-	"MISC53",
-	"MISC54",
-	"MISC55",
-	"MISC56",
-	"MISC57",
-	"MISC58",
-	"MISC59",
-	"MISC60",
-	"MISC61",
-	"MISC62",
-	"MISC63",
-	"MISC64",
-	"MISC65",
-	"MISC66",
-	"MISC67",
-	"MISC68",
-	"MISC69",
-	"MISC70",
-	"MISC71",
-	"MISC72",
-	"MISC73",
-	"MISC74",
-	"MISC75",
-	"MISC76",
-	"MISC77",
-	"MISC78",
-	"MISC79",
-	"MISC80",
-	"MISC81",
-	"MISC82",
-	"MISC83",
-	"MISC84",
-	"MISC85",
-	"MISC86",
-	"PUSH",
-	"PULL",
-	"DOGS",
-	"PLASMA1",
-	"PLASMA2"
-}
-
-SpriteNumber = {
---	"TROO",
-	"SHTG",
-	"PUNG",
-	"PISG",
-	"PISF",
-	"SHTF",
-	"SHT2",
-	"CHGG",
-	"CHGF",
-	"MISG",
-	"MISF",
-	"SAWG",
-	"PLSG",
-	"PLSF",
-	"BFGG",
-	"BFGF",
-	"BLUD",
-	"PUFF",
-	"BAL1",
-	"BAL2",
-	"PLSS",
-	"PLSE",
-	"MISL",
-	"BFS1",
-	"BFE1",
-	"BFE2",
-	"TFOG",
-	"IFOG",
-	"PLAY",
-	"POSS",
-	"SPOS",
-	"VILE",
-	"FIRE",
-	"FATB",
-	"FBXP",
-	"SKEL",
-	"MANF",
-	"FATT",
-	"CPOS",
-	"SARG",
-	"HEAD",
-	"BAL7",
-	"BOSS",
-	"BOS2",
-	"SKUL",
-	"SPID",
-	"BSPI",
-	"APLS",
-	"APBX",
-	"CYBR",
-	"PAIN",
-	"SSWV",
-	"KEEN",
-	"BBRN",
-	"BOSF",
-	"ARM1",
-	"ARM2",
-	"BAR1",
-	"BEXP",
-	"FCAN",
-	"BON1",
-	"BON2",
-	"BKEY",
-	"RKEY",
-	"YKEY",
-	"BSKU",
-	"RSKU",
-	"YSKU",
-	"STIM",
-	"MEDI",
-	"SOUL",
-	"PINV",
-	"PSTR",
-	"PINS",
-	"MEGA",
-	"SUIT",
-	"PMAP",
-	"PVIS",
-	"CLIP",
-	"AMMO",
-	"ROCK",
-	"BROK",
-	"CELL",
-	"CELP",
-	"SHEL",
-	"SBOX",
-	"BPAK",
-	"BFUG",
-	"MGUN",
-	"CSAW",
-	"LAUN",
-	"PLAS",
-	"SHOT",
-	"SGN2",
-	"COLU",
-	"SMT2",
-	"GOR1",
-	"POL2",
-	"POL5",
-	"POL4",
-	"POL3",
-	"POL1",
-	"POL6",
-	"GOR2",
-	"GOR3",
-	"GOR4",
-	"GOR5",
-	"SMIT",
-	"COL1",
-	"COL2",
-	"COL3",
-	"COL4",
-	"CAND",
-	"CBRA",
-	"COL6",
-	"TRE1",
-	"TRE2",
-	"ELEC",
-	"CEYE",
-	"FSKU",
-	"COL5",
-	"TBLU",
-	"TGRN",
-	"TRED",
-	"SMBT",
-	"SMGT",
-	"SMRT",
-	"HDB1",
-	"HDB2",
-	"HDB3",
-	"HDB4",
-	"HDB5",
-	"HDB6",
-	"POB1",
-	"POB2",
-	"BRS1",
-	"TLMP",
-	"TLP2",
-	"TNT1",
-	"DOGS",
-	"PLS1",
-	"PLS2",
-	"BON3",
-	"BON4",
-	"BLD2"
-}
 
 while true do
 	if Init then init_objects() end
